Support delay for new added nodes
New nodes could be in not Ready status for a while for initialization,
now with the new config option 'repair-delay-after-add', user can
set a duration, e.g. 10mins to let the auto healer skip the health
check for those new added nodes if there are added in 10mins. The
default value is 20mins.